requirements.txt
annotated-types==0.7.0
asttokens==2.4.1
backcall==0.2.0
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
executing==2.0.1
funkify==0.4.5
fusepy==3.0.1
imageio==2.34.2
importlib_metadata==8.2.0
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.1
jsonpickle==3.2.2
jupyter_client==8.6.2
jupyter_core==5.7.2
lazy_loader==0.4
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkg_resources==0.0.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyeditdistance==1.0.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
PyWavelets==1.4.1
pyzmq==26.1.0
requires==0.10.5
scikit-image==0.21.0
scipy==1.10.1
six==1.16.0
stack-data==0.6.3
tifffile==2023.7.10
tiledb==0.31.1
tiledb-bioimg==0.3.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
xtyping==0.8.2
zipp==3.20.0
